requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==4.2.21
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-xml==2.0.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.2.0
